+++ src/MentorLake.Gtk3/GLib/GIOCondition.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GLib;
 
 [Flags]
-public enum GIOCondition : long
+public enum GIOCondition : uint
 {
 	G_IO_IN = 1,
 	G_IO_OUT = 4,
