+++ src/MentorLake.Gtk3/GLib/GKeyFileFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GLib;
 
 [Flags]
-public enum GKeyFileFlags : long
+public enum GKeyFileFlags : uint
 {
 	G_KEY_FILE_NONE = 0,
 	G_KEY_FILE_KEEP_COMMENTS = 1,
