+++ src/MentorLake.Gtk3/GLib/GMarkupParseFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GLib;
 
 [Flags]
-public enum GMarkupParseFlags : long
+public enum GMarkupParseFlags : uint
 {
 	G_MARKUP_DEFAULT_FLAGS = 0,
 	G_MARKUP_DO_NOT_USE_THIS_UNSUPPORTED_FLAG = 1,
