+++ src/MentorLake.Gtk3/GModule/GModuleFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GModule;
 
 [Flags]
-public enum GModuleFlags : long
+public enum GModuleFlags : uint
 {
 	G_MODULE_BIND_LAZY = 1,
 	G_MODULE_BIND_LOCAL = 2,
