+++ src/MentorLake.Gtk3/GObject/GBindingFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GObject;
 
 [Flags]
-public enum GBindingFlags : long
+public enum GBindingFlags : uint
 {
 	G_BINDING_DEFAULT = 0,
 	G_BINDING_BIDIRECTIONAL = 1,
