+++ src/MentorLake.Gtk3/GObject/GConnectFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GObject;
 
 [Flags]
-public enum GConnectFlags : long
+public enum GConnectFlags : uint
 {
 	G_CONNECT_DEFAULT = 0,
 	G_CONNECT_AFTER = 1,
