+++ src/MentorLake.Gtk3/Gio/GBusNameWatcherFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GBusNameWatcherFlags : long
+public enum GBusNameWatcherFlags : uint
 {
 	G_BUS_NAME_WATCHER_FLAGS_NONE = 0,
 	G_BUS_NAME_WATCHER_FLAGS_AUTO_START = 1
