+++ src/MentorLake.Gtk3/Gio/GConverterFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GConverterFlags : long
+public enum GConverterFlags : uint
 {
 	G_CONVERTER_NO_FLAGS = 0,
 	G_CONVERTER_INPUT_AT_END = 1,
