+++ src/MentorLake.Gtk3/Gio/GDBusPropertyInfoFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusPropertyInfoFlags : long
+public enum GDBusPropertyInfoFlags : uint
 {
 	G_DBUS_PROPERTY_INFO_FLAGS_NONE = 0,
 	G_DBUS_PROPERTY_INFO_FLAGS_READABLE = 1,
