+++ src/MentorLake.Gtk3/Gio/GDBusSubtreeFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusSubtreeFlags : long
+public enum GDBusSubtreeFlags : uint
 {
 	G_DBUS_SUBTREE_FLAGS_NONE = 0,
 	G_DBUS_SUBTREE_FLAGS_DISPATCH_TO_UNENUMERATED_NODES = 1
