+++ src/MentorLake.Gtk3/Gio/GDriveStartFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDriveStartFlags : long
+public enum GDriveStartFlags : uint
 {
 	G_DRIVE_START_NONE = 0
 }
