+++ src/MentorLake.Gtk3/Gio/GFileMonitorFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GFileMonitorFlags : long
+public enum GFileMonitorFlags : uint
 {
 	G_FILE_MONITOR_NONE = 0,
 	G_FILE_MONITOR_WATCH_MOUNTS = 1,
