+++ src/MentorLake.Gtk3/Gio/GFileQueryInfoFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GFileQueryInfoFlags : long
+public enum GFileQueryInfoFlags : uint
 {
 	G_FILE_QUERY_INFO_NONE = 0,
 	G_FILE_QUERY_INFO_NOFOLLOW_SYMLINKS = 1
