+++ src/MentorLake.Gtk3/Gio/GMountMountFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GMountMountFlags : long
+public enum GMountMountFlags : uint
 {
 	G_MOUNT_MOUNT_NONE = 0
 }
