+++ src/MentorLake.Gtk3/Gio/GResolverNameLookupFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GResolverNameLookupFlags : long
+public enum GResolverNameLookupFlags : uint
 {
 	G_RESOLVER_NAME_LOOKUP_FLAGS_DEFAULT = 0,
 	G_RESOLVER_NAME_LOOKUP_FLAGS_IPV4_ONLY = 1,
