+++ src/MentorLake.Gtk3/Gio/GResourceFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GResourceFlags : long
+public enum GResourceFlags : uint
 {
 	G_RESOURCE_FLAGS_NONE = 0,
 	G_RESOURCE_FLAGS_COMPRESSED = 1
