+++ src/MentorLake.Gtk3/Gio/GResourceLookupFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GResourceLookupFlags : long
+public enum GResourceLookupFlags : uint
 {
 	G_RESOURCE_LOOKUP_FLAGS_NONE = 0
 }
